fixed subscription responses
now they return the subscribe state. and removed logs

+++ database/queries/queries_subscriptions.js
@@ -13,6 +13,9 @@ async function getSubscription(owner_id, subject_id, type="all"){
 async function addSubscription(owner_id, subject_id, type="all"){
     try {
         const {data, error} = await supabase.from('subscriptions').insert([{owner_id, subject_id, type}]).select()
+        console.log('query add sub')
+        console.log(`data ${data}`)
+        console.log(`error ${error}`)
         return error ? error : data
     } catch (error){
         console.log('error ao add subscription', error)
@@ -22,7 +25,9 @@ async function addSubscription(owner_id, subject_id, type="all"){
 
 async function removeSubscription(owner_id, subject_id, type=undefined){
     try {
-        const {data, error} = supabase.from('subscriptions').delete().eq('subject_id', subject_id)
+        const {data, error} = await supabase.from('subscriptions').delete().eq('owner_id', owner_id).eq('subject_id', subject_id).select()
+        console.log(`data ${data}`)
+        console.log(`error ${error}`)
         return error ? error : data
     } catch (error){
         console.log('error ao remove subscription', error)
